ajuste

+++ ecommerce/store/models.py
@@ -12,7 +12,7 @@ class Cliente(models.Model):
 
 class Produto(models.Model):
     nome = models.CharField(max_length=200, null=True)
-    preco = models.FloatField()
+    preco = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
     imagem = models.ImageField(null=True, blank=True)
